requirements.txt
tinkoff-investments>=0.2.0b110
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
loguru>=0.5.3
